Refactor temples
