+++ client/src/lib/navigation/Stacks.tsx
@@ -6,7 +6,7 @@ import {useRecoilValue} from 'recoil';
 import {killSwitchFields} from '../killSwitch/state/state';
 import Tabs from './Tabs';
 import Breathing from '../../routes/Breathing/Breathing';
-import Video from '../../routes/Video/Video';
+import Temple from '../../routes/temple/Temple';
 
 const Stack = createNativeStackNavigator<ScreenProps>();
 
@@ -25,7 +25,7 @@ const Stacks = () => {
         <>
           <Stack.Screen name={NAVIGATORS.TABS} component={Tabs} />
           <Stack.Screen name={ROUTES.BREATHING} component={Breathing} />
-          <Stack.Screen name={ROUTES.VIDEO} component={Video} />
+          <Stack.Screen name={ROUTES.TEMPLE} component={Temple} />
         </>
       )}
     </Stack.Navigator>
