+++ client/src/routes/Temple/DailyProvider.tsx
