+++ client/src/routes/Temple/VideoLayout.tsx
