+++ client/src/routes/Temples/api/temples.ts
@@ -1,0 +1,10 @@
+import {TEMPLE_ENDPOINT} from 'config';
+import {Temple} from '../../../../../shared/src/types/Temple';
+
+export const fetchTemples = async (): Promise<Temple[]> => {
+  try {
+    return (await fetch(TEMPLE_ENDPOINT)).json();
+  } catch (cause) {
+    throw new Error('Could not fetch temples', {cause});
+  }
+};
