+++ client/src/routes/Temples/hooks/useTemples.spec.ts
@@ -1,8 +1,8 @@
 import {act, renderHook} from '@testing-library/react-hooks';
-import {RecoilRoot} from 'recoil';
-import firestore from '@react-native-firebase/firestore';
+import {RecoilRoot, useRecoilValue} from 'recoil';
 import useTemples from './useTemples';
 import fetchMock, {enableFetchMocks} from 'jest-fetch-mock';
+import {isLoadingAtom, templesAtom} from '../state/state';
 
 enableFetchMocks();
 
@@ -15,48 +15,14 @@ afterEach(() => {
 });
 
 describe('useTemples', () => {
-  describe('subscribeTemple', () => {
-    it('should subscribe to live session document', async () => {
-      const {result} = renderHook(() => useTemples(), {
-        wrapper: RecoilRoot,
-      });
-
-      act(() => {
-        result.current.subscribeTemple('temple-id');
-      });
-      expect(firestore().collection).toHaveBeenCalledWith('temples');
-
-      expect(firestore().collection('temples').doc).toHaveBeenCalledWith(
-        'temple-id',
-      );
-      expect(
-        firestore().collection('temples').doc('temple-id').onSnapshot,
-      ).toHaveBeenCalled();
-    });
-
-    it('should set live content state', () => {
-      const {result} = renderHook(() => useTemples(), {wrapper: RecoilRoot});
-
-      act(() => {
-        result.current.subscribeTemple('test-id');
-      });
-
-      expect(result.current.contentState).toEqual({id: 'test-id'});
-    });
-
-    it('should return unsubscribe method when subscribing', () => {
-      const {result} = renderHook(() => useTemples(), {
-        wrapper: RecoilRoot,
-      });
-
-      act(() => {
-        const unsubscribe = result.current.subscribeTemple('temple-id');
-        expect(unsubscribe()).toEqual('unsubscribe-mock');
-      });
-    });
-  });
-
   describe('fetchTemples', () => {
+    const useTestHook = () => {
+      const {fetchTemples} = useTemples();
+      const temples = useRecoilValue(templesAtom);
+      const isLoading = useRecoilValue(isLoadingAtom);
+
+      return {fetchTemples, temples, isLoading};
+    };
     it('should fetch temples', async () => {
       fetchMock.mockResponseOnce(
         JSON.stringify([
@@ -64,7 +30,7 @@ describe('useTemples', () => {
         ]),
         {status: 200},
       );
-      const {result} = renderHook(() => useTemples(), {wrapper: RecoilRoot});
+      const {result} = renderHook(() => useTestHook(), {wrapper: RecoilRoot});
 
       await act(async () => {
         await result.current.fetchTemples();
@@ -83,7 +49,7 @@ describe('useTemples', () => {
         ]),
         {status: 200},
       );
-      const {result} = renderHook(() => useTemples(), {wrapper: RecoilRoot});
+      const {result} = renderHook(() => useTestHook(), {wrapper: RecoilRoot});
 
       const fetchPromise = act(async () => {
         await result.current.fetchTemples();
