--- functions/src/call/index.ts
@@ -1,17 +1,0 @@
-import 'firebase-functions';
-import {onRequest} from 'firebase-functions/v2/https';
-
-export const call = onRequest(
-  {
-    region: 'europe-west1',
-    memory: '256MiB',
-    maxInstances: 1024,
-    minInstances: 1,
-  },
-  async (request, response) => {
-    response
-      .status(200)
-      .send('https://29k-testing.daily.co/I1s53jXePycRMHTAwcQC');
-    return;
-  },
-);
