+++ functions/src/index.ts
@@ -2,5 +2,4 @@ import * as admin from 'firebase-admin';
 admin.initializeApp();
 
 export * from './killswitch';
-export * from './call';
 export * from './temple';
