+++ shared/src/types/Temple.ts
@@ -1,0 +1,6 @@
+export type Temple = {
+  id: string;
+  name: string;
+  url: string;
+  active: boolean;
+};
